fixed up tests, improved formatting for long charges for email, improved formatting for dob for emails

+++ tests/modules/test_export/test_save_html_payload.py
@@ -16,12 +16,12 @@ from modules.export import (
     convert_df_to_html,
 )
 from logs.config.logging import logs_config
-from locations import paths, dirs, root_dir
+from locations import paths, dirs, root_dir, test_dir
 
 
 mock_dirs = {
-    "payload_csv": Path("../../output/csv_converted_from_df/"),
-    "payload_email": Path("../../output/payload_email/"),
+    "payload_csv": test_dir / "output/csv_converted_from_df/",
+    "payload_email": test_dir / "output/payload_email/",
     "email_template": dirs["email_template"],  # using actual directory
 }
 
@@ -37,37 +37,31 @@ class TestSaveHtmlPayload(unittest.TestCase):
         logs_config(paths["logs_config_test"])
         # delete previous test output files
         logging.info(f"Deleting {mock_dirs['payload_email']} if it exists")
-        rmtree(mock_dirs["payload_email"])
+        if mock_dirs["payload_email"].is_dir():
+            rmtree(mock_dirs["payload_email"])
         # rebuild directory
         mock_dirs["payload_email"].mkdir(parents=True, exist_ok=True)
-        # create testing df
-        # self.df = convert_dict_into_df(docket_list, "Dauphin")
 
         # SET PANDAS OPTIONS FOR PRINT DISPLAY
         pd.set_option("display.max_columns", 20)
         pd.set_option("display.width", 2000)
         pd.set_option("display.max_rows", 700)
-        # # create testing df from problematic csv
-        # self.df = pd.read_csv(
-        # root_dir / "tests/fixtures/payload_csv/harrisburg_error.csv")
 
-        # GET PROBLEMATIC DF
-        self.df = pd.read_pickle(root_dir / "tests/fixtures/df_pkl/df.pkl")
-
-    # def tearDown(self) -> None:
-    #     print(f"Deleting temp folder: {mock_dirs['payload_email']}")
-    #     rmtree(mock_dirs["payload_email"])
 
     @mock.patch.dict(paths, mock_paths, clear=True)
     def test_html_file_is_created(self):
         """
         Test that an HTML file is generated
         """
-        styled_df = convert_df_to_html(self.df)
-        # save_html_county_payload("This is an introduction for the email",
-        #                          styled_df)
-        # # Check html file has been created
-        # self.assertTrue(mock_paths["payload_email"].is_file())
+        # create df
+        df = convert_dict_into_df(docket_list, "Dauphin")
+        # create styled df
+        styled_df = convert_df_to_html(df)
+        # wrap styled df with more html
+        save_html_county_payload("This is an introduction for the email",
+                                  styled_df)
+        # Check html file has been created
+        self.assertTrue(mock_paths["payload_email"].is_file())
 
 
 if __name__ == "__main__":
